requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
lxml==5.4.0
platformdirs==4.3.8
pydantic==2.11.5
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
zeep==4.3.1
